Update to 2.3.0

+++ Sources/API/Core/XML/dom_named_node_map.h
@@ -69,7 +69,7 @@ public:
 
 public:
 	/// \brief The number of nodes in the map.
-	int get_length() const;
+	unsigned long get_length() const;
 
 /// \}
 /// \name Operations
